+++ src/main/java/org/tingyu/atomicarchitect/common/util/math/Coordinate2i.java
@@ -1,0 +1,11 @@
+package org.tingyu.atomicarchitect.common.util.math;
+
+public class Coordinate2i {
+    public int x;
+    public int y;
+
+    public Coordinate2i(int x, int y) {
+        this.x = x;
+        this.y = y;
+    }
+}