+++ src/main/java/org/tingyu/atomicarchitect/common/world/level/block/entity/ModBlockEntityType.java
@@ -1,7 +1,0 @@
-package org.tingyu.atomicarchitect.common.world.level.block.entity;
-
-import net.minecraft.world.level.block.entity.BlockEntityType;
-
-public class ModBlockEntityType {
-    public static final BlockEntityType<AlchemyTableBlockEntity> ALCHEMY_TABLE_BLOCK_ENTITY = null;
-}